requirements.txt
asgiref==3.5.2
certifi==2022.9.24
click==8.0.4
dj-database-url==1.0.0
Django==4.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
whitenoise==6.2.0
psycopg2-binary==2.9.4
django_encrypted_files==0.0.10
PyExifTool==0.5.4